fix: shorter file input text
