+++ packages/common-helpers/src/locale/es.ts
@@ -64,12 +64,12 @@ export const localeInput: tLocaleInput = {
 	select_restablish_field: "Restablecer campo",
 	select_filter_options: "Escribe o haz doble click para las opciones",
 	file_one_of_amount: "{count} de {amount}",
-	file_delete_files: "Eliminar archivo | Eliminar archivos",
+	file_delete_files: "Quitar archivo | Quitar archivos",
 	file_thumb: "Miniatura",
-	file_choose_file: "Seleccionar un archivo | Seleccionar archivos",
+	file_choose_file: "Elegir archivo | Elegir archivos",
 	file_or_drop_files_here: "O arrástralo aquí | O arrástralos aquí",
 	file_max_file_size_mb: "máximo {size}MB por archivo",
-	file_drop_files_here: "Suelta el archivo aquí | Suelta los archivos aquí",
+	file_drop_files_here: "Soltar archivo aquí | Soltar archivos aquí",
 	file_completed: "Completado",
 	file_loading_files: "Cargando archivo... | Cargando archivos...",
 	swal: {
